Multilingual RAG System/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
google-generativeai
python-multipart
pydantic
PyMuPDF
